tests/wpt/tests/css/css-flexbox/parsing/flex-shrink-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Flexible Box Layout Module Level 1: parsing flex-shrink with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#propdef-flex-shrink">
<meta name="assert" content="flex-shrink supports only the grammar '<number>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("flex-shrink", "2e3.4");
test_invalid_value("flex-shrink", "-+5");
test_invalid_value("flex-shrink", "6 7");

test_invalid_value("flex-shrink", "1."); // Accepted by Edge.
</script>
</body>
</html>
